move data handle ctor tests to new file in unit_tests/view
