+++ core/unit_test/CMakeLists.txt
@@ -333,6 +333,9 @@ foreach(Tag Threads;Serial;OpenMP;Cuda;HPX;OpenMPTarget;OpenACC;HIP;SYCL)
              ViewCustomizationAccessorFromMapping
         )
       endif()
+      if(NOT Kokkos_ENABLE_IMPL_VIEW_LEGACY)
+        list(APPEND BV_TestNames ViewCtorDataHandle)
+      endif()
       set(${Tag}_VIEWSUPPORT)
       foreach(Name IN LISTS BV_TestNames)
         set(file ${dir}/Test${Tag}_View_${Name}.cpp)
