+++ core/unit_test/TestViewAPI.hpp
@@ -1571,40 +1571,6 @@ class TestViewAPI {
     typename multivector_type::const_type cmvX(cmv);
     typename const_multivector_type::const_type ccmvX(cmv);
   }
-
-  static void test_view_data_handle_ctor_ref_counts() {
-#ifndef KOKKOS_ENABLE_IMPL_VIEW_LEGACY
-    Kokkos::View<int *> a("A", 5);
-    ASSERT_EQ(a.use_count(), 1);
-    {
-      Kokkos::View<int *> b(a.data_handle(), a.extents());
-      ASSERT_EQ(a.use_count(), 2);
-      ASSERT_EQ(b.use_count(), 2);
-    }
-    ASSERT_EQ(a.use_count(), 1);
-    {
-      Kokkos::View<int *> b(a.data_handle(), a.mapping());
-      ASSERT_EQ(a.use_count(), 2);
-      ASSERT_EQ(b.use_count(), 2);
-    }
-    ASSERT_EQ(a.use_count(), 1);
-    {
-      Kokkos::View<int *> b(a.data_handle(), a.mapping(), a.accessor());
-      ASSERT_EQ(a.use_count(), 2);
-      ASSERT_EQ(b.use_count(), 2);
-    }
-    ASSERT_EQ(a.use_count(), 1);
-
-    // This is not currently supported but may be added in the future
-    //{
-    // Kokkos::View<int *> b(a.data_handle(), 5);
-    // ASSERT_EQ(a.use_count(), 2);
-    // ASSERT_EQ(b.use_count(), 2);
-    //}
-    static_assert(!std::constructible_from<Kokkos::View<int *>,
-                                           decltype(a.data_handle()), int>);
-#endif
-  }
 };
 
 }  // namespace Test
